fix flaky tests
