+++ dev-packages/browser-integration-tests/suites/tracing/metrics/web-vitals-inp/template.html
@@ -5,8 +5,8 @@
   </head>
   <body>
     <div>Rendered Before Long Task</div>
-    <button data-test-id="slow-button">Slow</button>
-    <button data-test-id="not-so-slow-button" data-sentry-component="not-so-slow-button" data-sentry-element="not-so-slow-button">Not so slow</button>
+    <button data-test-id="slow-button" data-sentry-element="SlowButton">Slow</button>
+    <button data-test-id="not-so-slow-button" data-sentry-element="NotSoSlowButton">Not so slow</button>
     <button data-test-id="normal-button" data-sentry-element="NormalButton">Click Me</button>
   </body>
 </html>
