+++ dev-packages/browser-integration-tests/suites/tracing/metrics/web-vitals-inp/test.ts
@@ -1,4 +1,3 @@
-import type { Route } from '@playwright/test';
 import { expect } from '@playwright/test';
 import type { Event as SentryEvent, SpanJSON } from '@sentry/types';
 
@@ -16,7 +15,6 @@ sentryTest('should capture an INP click event span.', async ({ browserName, getL
     sentryTest.skip();
   }
 
-  await page.route('**/path/to/script.js', (route: Route) => route.fulfill({ path: `${__dirname}/assets/script.js` }));
   await page.route('https://dsn.ingest.sentry.io/**/*', route => {
     return route.fulfill({
       status: 200,
@@ -31,11 +29,13 @@ sentryTest('should capture an INP click event span.', async ({ browserName, getL
   await getFirstSentryEnvelopeRequest<SentryEvent>(page); // wait for page load
 
   const spanEnvelopesPromise = getMultipleSentryEnvelopeRequests<SpanJSON>(page, 1, {
-    // envelopeType: 'span', // todo: does not work with envelopeType
+    envelopeType: 'span',
   });
 
   await page.locator('[data-test-id=normal-button]').click();
   await page.locator('.clicked[data-test-id=normal-button]').isVisible();
+
+  await page.waitForTimeout(500);
 
   // Page hide to trigger INP
   await page.evaluate(() => {
@@ -47,7 +47,7 @@ sentryTest('should capture an INP click event span.', async ({ browserName, getL
 
   expect(spanEnvelopes).toHaveLength(1);
   expect(spanEnvelopes[0].op).toBe('ui.interaction.click');
-  expect(spanEnvelopes[0].description).toBe('body > button.clicked');
+  expect(spanEnvelopes[0].description).toBe('body > NormalButton');
   expect(spanEnvelopes[0].exclusive_time).toBeGreaterThan(0);
   expect(spanEnvelopes[0].measurements?.inp.value).toBeGreaterThan(0);
   expect(spanEnvelopes[0].measurements?.inp.unit).toBe('millisecond');
@@ -62,9 +62,6 @@ sentryTest(
       sentryTest.skip();
     }
 
-    await page.route('**/path/to/script.js', (route: Route) =>
-      route.fulfill({ path: `${__dirname}/assets/script.js` }),
-    );
     await page.route('https://dsn.ingest.sentry.io/**/*', route => {
       return route.fulfill({
         status: 200,
@@ -78,15 +75,19 @@ sentryTest(
     await page.goto(url);
     await getFirstSentryEnvelopeRequest<SentryEvent>(page);
 
-    const spanEnvelopesPromise = getMultipleSentryEnvelopeRequests<SpanJSON>(page, 1, {
-      // envelopeType: 'span', // todo: does not work with envelopeType
-    });
-
     await page.locator('[data-test-id=normal-button]').click();
     await page.locator('.clicked[data-test-id=normal-button]').isVisible();
 
+    await page.waitForTimeout(500);
+
     await page.locator('[data-test-id=slow-button]').click();
     await page.locator('.clicked[data-test-id=slow-button]').isVisible();
+
+    await page.waitForTimeout(500);
+
+    const spanEnvelopesPromise = getMultipleSentryEnvelopeRequests<SpanJSON>(page, 1, {
+      envelopeType: 'span',
+    });
 
     // Page hide to trigger INP
     await page.evaluate(() => {
@@ -98,7 +99,7 @@ sentryTest(
 
     expect(spanEnvelopes).toHaveLength(1);
     expect(spanEnvelopes[0].op).toBe('ui.interaction.click');
-    expect(spanEnvelopes[0].description).toBe('body > button.clicked');
+    expect(spanEnvelopes[0].description).toBe('body > SlowButton');
     expect(spanEnvelopes[0].exclusive_time).toBeGreaterThan(400);
     expect(spanEnvelopes[0].measurements?.inp.value).toBeGreaterThan(400);
     expect(spanEnvelopes[0].measurements?.inp.unit).toBe('millisecond');
